- Migrated to Mobile App Token API - Removed StealthConfig
