+++ xfinity-usage/config.yaml
@@ -1,7 +1,7 @@
 name: "Xfinity Internet Usage"
 description: "Get Xfinity Internet Usage Data"
 url: "https://github.com/thor0215/hassio-xfinity-usage"
-version: "0.1.1"
+version: "0.1.3"
 image: "ghcr.io/thor0215/hassio-xfinity-usage-{arch}"
 slug: "xfinity-usage"
 init: false
@@ -36,6 +36,8 @@ options:
 schema:
   xfinity_username: str
   xfinity_password: password
+  refresh_token: str(,644)?
+  client_secret: str?
   page_timeout: int(15,)?
   polling_rate: int(30,)?
   log_level: list(info|debug)?
@@ -45,5 +47,4 @@ schema:
   mqtt_username: str?
   mqtt_password: password?
   mqtt_raw_usage: bool?
-  profile_cleanup: bool?
   debug_support: bool?
