+++ xfinity-usage/requirements.txt
@@ -4,3 +4,4 @@ requests==2.32.3
 PyJWT==2.10.1
 paho-mqtt==2.1.0
 colorlog==6.9.0
+cryptography==44.0.0
